Fixed #849 PowerMock.replayAll does not handle mocks created by annotation processor for `org.easymock.Mock` annotation
