+++ tests/easymock/junit412/src/main/java/samples/junit412/bug/github668/IncidentPropertyChangeDAO.java
@@ -1,4 +1,4 @@
-package samples.junit412.github668;
+package samples.junit412.bug.github668;
 
 /**
  *
