+++ tests/easymock/junit412/src/test/java/samples/junit412/bug/github668/TwoMockFieldsWithDifferentTypesClass.java
@@ -1,4 +1,4 @@
-package samples.junit412.github668;
+package samples.junit412.bug.github668;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -6,6 +6,7 @@ import org.powermock.api.easymock.annotation.Mock;
 import org.powermock.api.extension.listener.AnnotationEnabler;
 import org.powermock.core.classloader.annotations.PowerMockListener;
 import org.powermock.modules.junit4.PowerMockRunner;
+import samples.Service;
 
 import static org.easymock.EasyMock.expect;
 import static org.junit.Assert.assertEquals;
@@ -13,33 +14,30 @@ import static org.junit.Assert.assertNotNull;
 import static org.powermock.api.easymock.PowerMock.replayAll;
 
 
-/**
- *
- */
 @RunWith(PowerMockRunner.class)
 @PowerMockListener(AnnotationEnabler.class)
-public class TwoMockFieldsWithSameTypeCase {
+public class TwoMockFieldsWithDifferentTypesClass {
 
-    @Mock(fieldName = "incidentPropertyChangeDAO")
+    @Mock
     private IncidentPropertyChangeDAO incidentPropertyChangeDAO;
 
-    @Mock(fieldName = "propertyChangeDAO")
-    private IncidentPropertyChangeDAO propertyChangeDAO;
+    @Mock
+    private Service serviceMock;
 
     @Test
     public void mockClassShouldInjected() {
         assertNotNull(incidentPropertyChangeDAO);
-        assertNotNull(propertyChangeDAO);
+        assertNotNull(serviceMock);
     }
     
     @Test
     public void shouldBeAbleMockMethodsOfInjected() {
         expect(incidentPropertyChangeDAO.getIncident()).andReturn("value");
-        expect(propertyChangeDAO.getIncident()).andReturn("value1");
+        expect(serviceMock.getServiceMessage()).andReturn("value");
 
-        replayAll(incidentPropertyChangeDAO,propertyChangeDAO);
+        replayAll(incidentPropertyChangeDAO, serviceMock);
 
         assertEquals("value", incidentPropertyChangeDAO.getIncident());
-        assertEquals("value1", propertyChangeDAO.getIncident());
+        assertEquals("value", serviceMock.getServiceMessage());
     }
 }
