+++ tests/easymock/junit412/src/test/java/samples/junit412/bug/github668/TwoMockFieldsWithSameTypeCase.java
@@ -1,4 +1,4 @@
-package samples.junit412.github668;
+package samples.junit412.bug.github668;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -6,7 +6,6 @@ import org.powermock.api.easymock.annotation.Mock;
 import org.powermock.api.extension.listener.AnnotationEnabler;
 import org.powermock.core.classloader.annotations.PowerMockListener;
 import org.powermock.modules.junit4.PowerMockRunner;
-import samples.Service;
 
 import static org.easymock.EasyMock.expect;
 import static org.junit.Assert.assertEquals;
@@ -19,28 +18,28 @@ import static org.powermock.api.easymock.PowerMock.replayAll;
  */
 @RunWith(PowerMockRunner.class)
 @PowerMockListener(AnnotationEnabler.class)
-public class TwoMockFieldsWithDifferentTypesClass {
+public class TwoMockFieldsWithSameTypeCase {
 
-    @Mock
+    @Mock(fieldName = "incidentPropertyChangeDAO")
     private IncidentPropertyChangeDAO incidentPropertyChangeDAO;
 
-    @Mock
-    private Service serviceMock;
+    @Mock(fieldName = "propertyChangeDAO")
+    private IncidentPropertyChangeDAO propertyChangeDAO;
 
     @Test
     public void mockClassShouldInjected() {
         assertNotNull(incidentPropertyChangeDAO);
-        assertNotNull(serviceMock);
+        assertNotNull(propertyChangeDAO);
     }
     
     @Test
     public void shouldBeAbleMockMethodsOfInjected() {
         expect(incidentPropertyChangeDAO.getIncident()).andReturn("value");
-        expect(serviceMock.getServiceMessage()).andReturn("value");
+        expect(propertyChangeDAO.getIncident()).andReturn("value1");
 
-        replayAll(incidentPropertyChangeDAO, serviceMock);
+        replayAll(incidentPropertyChangeDAO,propertyChangeDAO);
 
         assertEquals("value", incidentPropertyChangeDAO.getIncident());
-        assertEquals("value", serviceMock.getServiceMessage());
+        assertEquals("value1", propertyChangeDAO.getIncident());
     }
 }
