+++ tests/easymock/junit412/src/test/java/samples/junit412/bug/github668/package-info.java
@@ -2,4 +2,4 @@
  * \@Mock fields are not injected anymore (1.6.5 regression)
  * https://github.com/jayway/powermock/issues/668
  */
-package samples.junit412.github668;
+package samples.junit412.bug.github668;
