+++ tests/easymock/junit412/src/test/java/samples/junit412/bug/github849/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * PowerMock.replayAll does not handle mocks created by annotation processor for `org.easymock.Mock` annotation
+ */
+package samples.junit412.bug.github849;
